moved get user by id/update with id to api/user/userId folder; also renamed users.tsx to route.ts
