+++ src/app/api/users/route.ts
@@ -42,32 +42,7 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
         } catch (err) {
             res.status(500).json({ error: 'failed to load data' });
         }
-    } else if (req.method === 'PUT') {
-        try {
-            if (!req.body) {
-                return res.status(400).json({ error: 'missing request body' });
-            }
-
-            const userid = req.query.userid;
-            let info : UpdateData<User> = req.body;
-
-            if (!userid) {
-                return res.status(400).json({ error: 'missing userid' });
-            } else if (!info) {
-                return res.status(400).json({ error: 'invalid user data' });
-            } else {
-                try {
-                    let strUserId = String(userid);
-                    const result = await updateAccount(strUserId, info);
-                    res.status(200).json({ result });
-                } catch {
-                    res.status(404).json({ error: 'error with creating the account' });
-                }
-            }
-        } catch (err) {
-            res.status(500).json({ error: 'failed to load data' });
-        }
-    }
+    } 
 }
 
 function checkIfValidUser(user : User) {
